requirement.txt
langchain
langchain_community
langgraph
langserve[all]
python-dotenv
langchain_google_genai
langchain_core